+++ src/.idea/misc.xml
@@ -4,6 +4,7 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="..\:/Users/IraKu/Desktop/GenLib/soft_eng/src/app/src/main/res/layout/activity_main.xml" value="0.25" />
+        <entry key="..\:/Users/IraKu/Desktop/GenLib/soft_eng/src/app/src/main/res/layout/window_main.xml" value="0.27291666666666664" />
       </map>
     </option>
   </component>
