+++ src/app/src/main/res/values/strings.xml
@@ -4,4 +4,10 @@
     <string name="button_register">Регистрация</string>
     <string name="field_login">Логин</string>
     <string name="field_password">Пароль</string>
+    <string name="title_activity_main2">MainActivity2</string>
+    <!-- Strings used for fragments for navigation -->
+    <string name="first_fragment_label">First Fragment</string>
+    <string name="second_fragment_label">Second Fragment</string>
+    <string name="next">Next</string>
+    <string name="previous">Previous</string>
 </resources>
